refactor into js and mojo implementations
