+++ lib/src/js/util.dart
@@ -1,0 +1,10 @@
+library firebase.js.util;
+
+import 'dart:js';
+
+jsify(value) {
+  if (value is Map || value is Iterable) {
+    return new JsObject.jsify(value);
+  }
+  return value;
+}
